Access levels implemented, test project took out, will be replaced

+++ Infrastructure/ProductRepository.cs
@@ -22,22 +22,31 @@ public class ProductRepository : IProductRepository
         return _context.ProductTable.Find(id) ?? throw new KeyNotFoundException();
     }
 
-    public Product CreateNewProduct(Product product)
+    public Product CreateNewProduct(Product product, int userId)
     {
+        UserInWarehouse uiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+        if (uiw.AccessLevel != 4)
+            throw new Exception("Access Denied");
         _context.ProductTable.Add(product);
         _context.SaveChanges();
         return product;
     }
 
-    public Product UpdateProduct(Product product)
+    public Product UpdateProduct(Product product, int userId)
     {
+        UserInWarehouse uiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+        if (uiw.AccessLevel != 4)
+            throw new Exception("Access Denied");
         _context.ProductTable.Update(product);
         _context.SaveChanges();
         return product;
     }
 
-    public Product DeleteProduct(int id)
+    public Product DeleteProduct(int id, int userId)
     {
+        UserInWarehouse uiw = _context.UserWarehouseTable.Where(u => u.UserId == userId).First() ?? throw new KeyNotFoundException();
+        if (uiw.AccessLevel != 4)
+            throw new Exception("Access Denied");
         Product productToDelete = _context.ProductTable.Find(id) ?? throw new KeyNotFoundException();
         _context.ProductTable.Remove(productToDelete);
         _context.SaveChanges();
